+++ packages/manager/modules/exchange/src/dashboard/Exchange.class.js
