+++ packages/manager/modules/exchange/src/dashboard/exchange.tabs.controller.js
