--- packages/manager/modules/exchange/src/exchanges/translations/Messages_fr_FR.json
@@ -1,3 +1,0 @@
-{
-  "exchange_title": "Exchange"
-}
